+++ probatus/binning/__init__.py
@@ -1,3 +1,3 @@
-from .binning import SimpleBucketer, AgglomerativeBucketer, QuantileBucketer
+from .binning import SimpleBucketer, AgglomerativeBucketer, QuantileBucketer, TreeBucketer
 
-__all__ = ['SimpleBucketer', 'AgglomerativeBucketer', 'QuantileBucketer']
+__all__ = ['SimpleBucketer', 'AgglomerativeBucketer', 'QuantileBucketer', 'TreeBucketer']