+++ probatus/utils/__init__.py
@@ -3,7 +3,8 @@ from .scoring import Scorer, get_scorers
 from .arrayfuncs import assure_numpy_array, assure_pandas_df, check_1d, warn_if_missing
 from .warnings import ApproximationWarning
 from ._utils import class_name_from_object, assure_list_of_strings, assure_list_values_allowed
+from .tree import TreePathFinder
 
 __all__ = ['NotFittedError', 'DimensionalityError', 'UnsupportedModelError', 'Scorer', 'assure_numpy_array', 'assure_pandas_df', 'check_1d',
            'ApproximationWarning', 'class_name_from_object', 'get_scorers', 'assure_list_of_strings',
-           'assure_list_values_allowed', 'warn_if_missing']
+           'assure_list_values_allowed', 'warn_if_missing','TreePathFinder']