+++ probatus/utils/warnings.py
@@ -1,3 +1,8 @@
 class ApproximationWarning(Warning):
     def __init__(self, message):
         self.message = message
+
+
+class NotIntendedUseWarning(Warning):
+    def __init__(self, message):
+        self.message = message