+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- TreeBucketer
+- TreePathFinder Helper class to find the boundaries of the leaves of the Decision Tree
 
 ## [1.0.0] - 2020-02-24
 - metric_volatility and sample_similarity rebuilt